move AutoSubmitFilterForm in form folder
