--- packages/ra-core/src/controller/list/filter/index.ts
@@ -1,1 +1,0 @@
-export * from './AutoSubmitFilterForm';
