+++ packages/ra-core/src/form/index.ts
@@ -48,3 +48,4 @@ export * from './useSuggestions';
 export * from './useUnique';
 export * from './useWarnWhenUnsavedChanges';
 export * from './WarnWhenUnsavedChanges';
+export * from './AutoSubmitFilterForm';
